requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
azure-cosmos>=4.7.0
aiohttp>=3.9.0
